+++ Kexa/helpers/jsonStringify.ts
@@ -1,0 +1,3 @@
+export function jsonStringify(data: any, space?: number) {
+    return JSON.stringify(data, (_, v) => typeof v === 'bigint' ? v.toString() : v, space);
+}