+++ Kexa/services/addOn/display/kubernetesDisplay.service.ts
@@ -16,6 +16,8 @@ export function propertyToSend(rule: Rules, objectContent: any, isSms: boolean=f
                 return `Name : ` + objectContent.metadata.name + `</br>NameSpace : ` + objectContent.metadata.namespace + `</br>Scanned since ` + objectContent.interval
         case "helm":
             return `Helm name : ` + objectContent?.metadata?.name + ` with uid : ` + objectContent?.metadata?.uid
+        case "podsConsumption":
+            return 'pod : ' + objectContent?.podName + ' in NameSpace : ' + objectContent?.metadata?.namespace;
         default:
             return 'resource : Id : ' + objectContent?.metadata?.name + ' in NameSpace : ' + objectContent?.metadata?.namespace;
     }