+++ Kexa/services/addOn/gcpGathering.service.ts
@@ -34,6 +34,7 @@ import { GCPResources } from "../../models/gcp/resource.models";
 import {Storage} from "@google-cloud/storage";
 import { deleteFile, getFile, writeStringToJsonFile } from "../../helpers/files";
 import { GcpConfig } from "../../models/gcp/config.models";
+import { jsonStringify } from '../../helpers/jsonStringify';
 
 ////////////////////////////////
 //////   INITIALIZATION   //////
@@ -271,14 +272,14 @@ async function executeAllRegions(projectId: number, serviceFunction: Function, c
             const request = { parent };
             if (!isIterable) {
                 let response = await serviceFunction.call(client, request);
-                let jsonTmp = JSON.parse(JSON.stringify(response));
+                let jsonTmp = JSON.parse(jsonStringify(response));
                 if (jsonTmp != null && jsonTmp.length > 0) {
                     jsonResponses.push(addRegionGCP(jsonTmp, currentRegion));
                 }
             } else {
                 const iterable = await serviceFunction.call(client, request);
                 for await (const response of iterable) {
-                    let jsonTmp = JSON.parse(JSON.stringify(response));
+                    let jsonTmp = JSON.parse(jsonStringify(response));
                     if (jsonTmp != null && jsonTmp.length > 0) {
                         jsonResponses.push(addRegionGCP(jsonTmp, currentRegion));
                     }
@@ -338,7 +339,7 @@ async function listAllComputes(projectId: string): Promise<Array<any>|null> {
 
         if (instances && instances.length > 0) {
             for (let i = 0; i < instances.length; i++) {
-                jsonData.push(JSON.parse(JSON.stringify(instances[i].metadata.items.length)));
+                jsonData.push(JSON.parse(jsonStringify(instances[i].metadata.items.length)));
             }
         }
     }
@@ -358,7 +359,7 @@ async function listSSHKey(projectId: string): Promise<Array<any>|null> {
         const instances = instancesObject.instances;
         if (instances && instances.length > 0) {
             for (let i = 0; i < instances.length; i++) {
-                jsonData.push(JSON.parse(JSON.stringify(instances[i].metadata?.items)));
+                jsonData.push(JSON.parse(jsonStringify(instances[i].metadata?.items)));
             }
         }
     }
@@ -376,7 +377,7 @@ async function listPersistentDisks(projectId: any) {
 
         if (disks && disks.length > 0) {
             for (let i = 0; i < disks.length; i++) {
-                jsonData.push(JSON.parse(JSON.stringify(disks[i])));
+                jsonData.push(JSON.parse(jsonStringify(disks[i])));
             }
         }
     }
@@ -392,7 +393,7 @@ async function listAllBucket(): Promise<Array<any>|null> {
         const [buckets] = await storage.getBuckets();
         for (let i = 0; i < buckets.length; i++) {
             const current_bucket = await storage.bucket(buckets[i].name).get();
-            const jsonData = JSON.parse(JSON.stringify(current_bucket));
+            const jsonData = JSON.parse(jsonStringify(current_bucket));
             jsonReturn.push(jsonData);
 
         }
@@ -420,7 +421,7 @@ async function listAllClusters(): Promise<Array<any>|null> {
             zone: "-",
         };
         const [response] = await cnt.listClusters(request);
-        jsonData = JSON.parse(JSON.stringify(response));
+        jsonData = JSON.parse(jsonStringify(response));
     } catch (e) {
         logger.debug(e);
     }
@@ -437,7 +438,7 @@ async function listAllProject(): Promise<Array<any>|null> {
         const client = new ProjectsClient();
         const projects = client.searchProjectsAsync();
         for await (const project of projects) {
-            jsonData = JSON.parse(JSON.stringify(project));
+            jsonData = JSON.parse(jsonStringify(project));
         }
     } catch (e) {
         logger.debug(e);
@@ -457,7 +458,7 @@ async function getBillingAccount(projectId: any): Promise<Array<any>|null> {
             projectId,
         });
         for (const account of accounts) {
-            jsonData = JSON.parse(JSON.stringify(account));
+            jsonData = JSON.parse(jsonStringify(account));
         }
     } catch (e) {
         logger.debug(e);
@@ -476,7 +477,7 @@ async function listWorkflows(projectId: any): Promise<Array<any>|null> {
             parent: client.locationPath(projectId, '-'),
         });
         for (const workflow of workflows) {
-            jsonData = JSON.parse(JSON.stringify(workflow));
+            jsonData = JSON.parse(jsonStringify(workflow));
         }
     } catch (e) {
         logger.debug(e);
@@ -494,7 +495,7 @@ async function listWebSecurityConfig(projectId: any): Promise<Array<any>|null> {
         const stats = await client.listScanConfigs({
             parent: `projects/${projectId}`,
         });
-        jsonData = JSON.parse(JSON.stringify(stats));
+        jsonData = JSON.parse(jsonStringify(stats));
     }
     catch (e) {
         logger.debug(e);
@@ -531,7 +532,7 @@ async function listVMWareEngine(projectId: any): Promise<Array<any>|null>  {
             request
         );
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.debug(e);
@@ -567,7 +568,7 @@ export async function listSecrets(projectId: any): Promise<Array<any>|null> {
         const request = { parent };
         const iterable = await secretmanagerClient.listSecretsAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.debug(e);
@@ -586,7 +587,7 @@ async function listConnectivityTests(projectId: any): Promise<Array<any>|null>  
         const tests = await client.listConnectivityTests({
             parent: `projects/${projectId}/locations/global`,
         });
-        jsonData = JSON.parse(JSON.stringify(tests));
+        jsonData = JSON.parse(jsonStringify(tests));
     } catch (e) {
         logger.debug(e);
     }
@@ -605,7 +606,7 @@ async function listResourceSettings(projectId: any): Promise<Array<any>|null> {
         const settings = await client.listSettings({
             parent: `projects/${projectId}`,
         });
-        jsonData = JSON.parse(JSON.stringify(settings));
+        jsonData = JSON.parse(jsonStringify(settings));
     } catch (e) {
         logger.debug(e);
     }
@@ -638,7 +639,7 @@ async function listOSConfig(projectId: any): Promise<Array<any>|null> {
         const [deployments] = await client.listPatchDeployments({
             parent: `projects/${projectId}`,
         });
-        jsonData = JSON.parse(JSON.stringify(deployments));
+        jsonData = JSON.parse(jsonStringify(deployments));
     } catch (e) {
         logger.debug(e);
     }
@@ -656,7 +657,7 @@ async function listOrgPolicyContraints(projectId: any): Promise<Array<any>|null>
         const constraints = await client.listConstraints({
             parent: `projects/${projectId}`,
         });
-        jsonData = JSON.parse(JSON.stringify(constraints));
+        jsonData = JSON.parse(jsonStringify(constraints));
     } catch (e) {
         logger.debug(e);
     }
@@ -690,7 +691,7 @@ async function listNotebookInstances(projectId: any): Promise<Array<any> | null>
             parent: `projects/${projectId}/locations/-`,
         });
         for (const instance of instances) {
-            jsonData = JSON.parse(JSON.stringify(instance));
+            jsonData = JSON.parse(jsonStringify(instance));
         }
     } catch (e) {
         logger.debug(e);
@@ -709,7 +710,7 @@ async function listDashboards(projectId: any): Promise<Array<any> | null> {
         const ds = new DashboardsServiceClient();
         const [dashboards] = await ds.listDashboards({parent,});
         for (const dashboard of dashboards) {
-            jsonData = JSON.parse(JSON.stringify(dashboard));
+            jsonData = JSON.parse(jsonStringify(dashboard));
         }
     } catch (e) {
         logger.debug(e);
@@ -728,7 +729,7 @@ async function listIdentitiesDomain(projectId: any): Promise<Array<any> | null> 
         const domains = await client.listDomains({
             parent: `projects/${projectId}/locations/global`,
         });
-        jsonData = JSON.parse(JSON.stringify(domains));
+        jsonData = JSON.parse(jsonStringify(domains));
     } catch (e) {
         logger.debug(e)
     }
@@ -746,7 +747,7 @@ async function listLineageProcesses(projectId: any): Promise<Array<any> | null> 
         const request = {parent};
         const iterable = await lineageClient.listProcessesAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.debug(e);
@@ -768,7 +769,7 @@ async function listKMSCryptoKeys(projectId: any): Promise<Array<any> | null> {
             maxResults: 50,
             autoPaginate: false,
         });
-        jsonData = JSON.parse(JSON.stringify(response));
+        jsonData = JSON.parse(jsonStringify(response));
     } catch (e) {
         logger.debug(e);
     }
@@ -788,7 +789,7 @@ async function listKMSKeyRings(projectId: any): Promise<Array<any> | null> {
             parent: locationName
         });
         for (const keyRing of keyRings) {
-            jsonData = JSON.parse(JSON.stringify(keyRing));
+            jsonData = JSON.parse(jsonStringify(keyRing));
         }
     } catch (e) {
         logger.debug(e);
@@ -807,7 +808,7 @@ async function listDomainsRegistration(projectId: any): Promise<Array<any> | nul
         const [registrations] = await client.listRegistrations({
             parent: `projects/${projectId}/locations/global`,
         });
-        jsonData = JSON.parse(JSON.stringify(registrations));
+        jsonData = JSON.parse(jsonStringify(registrations));
     } catch (e) {
         logger.debug(e);
     }
@@ -824,7 +825,7 @@ async function listDnsZones(projectId: any): Promise<Array<any> | null> {
         const dns = new DNS();
         const [zones] = await dns.getZones();
         for (const zone of zones) {
-            jsonData = JSON.parse(JSON.stringify(zone));
+            jsonData = JSON.parse(jsonStringify(zone));
         }
     } catch (e) {
         logger.debug(e);
@@ -859,7 +860,7 @@ async function listCertificates(projectId: any): Promise<Array<any> | null> {
         const request = { parent };
         const iterable = await certificatemanagerClient.listCertificatesAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.debug(e);
@@ -897,7 +898,7 @@ async function listWorkloads(projectId: any): Promise<Array<any> | null> {
             parent: `organizations/${projectId}`,
         });
         for (const workload of workloads) {
-            jsonData = JSON.parse(JSON.stringify(workload));
+            jsonData = JSON.parse(jsonStringify(workload));
         }*/
     } catch (e) {
         logger.debug(e);
@@ -948,11 +949,11 @@ async function listTagsKeys(projectId: any, regionsList: Array<string>): Promise
         const rep = await tagsValueClient.listTagValues();
         console.log("reponse here");
         console.log(rep);
-        jsonData = JSON.parse(JSON.stringify(rep));
+        jsonData = JSON.parse(jsonStringify(rep));
        /* const iterable = await tagsKeyClient.listTagKeysAsync();
         for await (const response of iterable) {
             console.log(response);
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }*/
     } catch (e) {
         console.log("error tags:",e);
@@ -977,7 +978,7 @@ async function listAppConnectors(projectId: any): Promise<Array<any> | null> {
         const request = {parent};
         const iterable = await appconnectorsClient.listAppConnectorsAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.error(e);
@@ -998,7 +999,7 @@ async function listApiKeys(projectId: any): Promise<Array<any> | null> {
         const request = { parent };
         const iterable = await apikeysClient.listKeysAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.error(e);
@@ -1018,7 +1019,7 @@ async function listApi(projectId: any): Promise<Array<any> | null> {
             parent: `projects/${projectId}/locations/global`,
         });
         for (const api of apis) {
-            jsonData = JSON.parse(JSON.stringify(api));
+            jsonData = JSON.parse(jsonStringify(api));
         }
     } catch (e) {
         logger.error(e);
@@ -1037,7 +1038,7 @@ async function listAccessPolicy(projectId: any): Promise<Array<any> | null> {
         const requests = await client.listApprovalRequests({
             parent: `projects/${projectId}`,
         });
-        jsonData = JSON.parse(JSON.stringify(requests));
+        jsonData = JSON.parse(jsonStringify(requests));
     } catch (e) {
         logger.error(e);
     }
@@ -1080,7 +1081,7 @@ async function listStorageConfig(projectId: any): Promise<Array<any>|null> {
         };
         const iterable = await storageinsightsClient.listReportConfigsAsync(request);
         for await (const response of iterable) {
-            jsonData = JSON.parse(JSON.stringify(response));
+            jsonData = JSON.parse(jsonStringify(response));
         }
     } catch (e) {
         logger.error(e);
@@ -1102,7 +1103,7 @@ async function listPrivateCertificates(projectId: any): Promise<Array<any>|null>
         const res = await client.listCertificates({
             parent: `projects/${projectId}/locations/-/caPools/-`,
         });
-        jsonData = JSON.parse(JSON.stringify(res));
+        jsonData = JSON.parse(jsonStringify(res));
     } catch (e) {
         logger.error(e);
     }