+++ Kexa/services/analyse.service.ts
@@ -25,6 +25,8 @@ import {getContext, getNewLogger} from "./logger.service";
 import { splitProperty } from '../helpers/spliter';
 import { downloadFile, unzipFile } from '../helpers/dowloadFile';
 import { getConfig } from '../helpers/loaderConfig';
+import { jsonStringify } from '../helpers/jsonStringify';
+import { Memoisation } from './memoisation.service';
 const logger = getNewLogger("AnalyseLogger");
 
 const jsome = require('jsome');
@@ -93,7 +95,7 @@ export function extractAddOnNeed(settingFileList: SettingFile[]){
             if(!objectNameList[ruleFile.alert.global.name][rule.cloudProvider].includes(rule.objectName)) objectNameList[ruleFile.alert.global.name][rule.cloudProvider].push(rule.objectName);
         });
     });
-    writeStringToJsonFile(JSON.stringify({ "addOn" : providerList, "objectNameNeed": objectNameList }), "./config/addOnNeed.json");
+    writeStringToJsonFile(jsonStringify({ "addOn" : providerList, "objectNameNeed": objectNameList }), "./config/addOnNeed.json");
 }
 
 function getListNeedRules(): string[]{
@@ -435,10 +437,11 @@ export function checkRules(rules:Rules[], resources:ProviderResource, alert: Ale
     return result;
 }
 function actionAfterCheckRule(rule: Rules, objectResource: any, alert: Alert, subResultScan: SubResultScan[]): SubResultScan[] {
-    logger.debug("subResultScan:"+JSON.stringify(subResultScan));
     let error = subResultScan.filter((value) => !value.result);
     if(error.length > 0){
-        alertFromRule(rule, subResultScan, objectResource, alert);
+        if (Memoisation.needToBeCache(rule, objectResource, rule.cloudProvider)){
+            alertFromRule(rule, subResultScan, objectResource, alert);
+        }
     }
     return error;
 }
@@ -499,7 +502,7 @@ export function parentResultScan(subResultScans: SubResultScan[], result: boolea
 export function checkCondition(condition:RulesConditions, resource:any): SubResultScan {
     try{
         let value = getSubProperty(resource, condition.property);
-        if (!condition.value)
+        if (condition.value === undefined)
             condition.value =  '';
         if (value === undefined)
             value = '';
@@ -611,13 +614,15 @@ export function checkEqual(condition:RulesConditions, value:any): boolean {
 
 export function checkGreaterThan(condition:RulesConditions, value:any): boolean {
     logger.debug("check greater than:" + value + " > " + condition.value + " ?");
-    if(value > condition.value) return true;
+    if(typeof value === "number" && value > (condition.value as number)) return true;
+    if(~~value > ~~condition.value) return true;
     return false;
 }
 
 export function checkLessThan(condition:RulesConditions, value:any): boolean {
     logger.debug("check less than:" + value + " < " + condition.value + " ?");
-    if(value < condition.value) return true;
+    if(typeof value === "number" && value < (condition.value as number)) return true;
+    if(~~value < ~~condition.value) return true;
     return false;
 }